test_simple_compute_shader
